feat(camera): deploy dumb camera

+++ raspberrypi_central/install.sh
@@ -48,10 +48,13 @@ fi
 
 source ~/.device
 
-# Change hostname "raspberrypi" to the device id, so it's easier to find the device on the network.
-CURRENT_HOSTNAME="raspberrypi"
-NEW_HOSTNAME=$DEVICE_ID
+grep -q "$DEVICE_ID" /etc/hosts
+if [ $? -eq 1 ]; then
+  # Change hostname "raspberrypi" to the device id, so it's easier to find the device on the network.
+  CURRENT_HOSTNAME="raspberrypi"
+  NEW_HOSTNAME=$DEVICE_ID
 
-sudo hostnamectl set-hostname "$NEW_HOSTNAME"
-sudo sed -i "s/127.0.1.1.*$CURRENT_HOSTNAME/127.0.1.1\t$NEW_HOSTNAME/g" /etc/hosts
-sudo reboot
+  sudo hostnamectl set-hostname "$NEW_HOSTNAME"
+  sudo sed -i "s/127.0.1.1.*$CURRENT_HOSTNAME/127.0.1.1\t$NEW_HOSTNAME/g" /etc/hosts
+  sudo reboot
+fi
